Moved NpgsqlDatabaseInfo to Internal
